pyproject.toml
[project]
name = "retrieval"
version = "0.1.0"
description = ""
authors = [
    {name = "SBSun",email = "byung0216@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi (>=0.115.6,<0.116.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "langchain-community (>=0.3.14,<0.4.0)",
    "langchain (>=0.3.14,<0.4.0)",
    "black (>=24.10.0,<25.0.0)",
    "langchain-ollama (>=0.2.2,<0.3.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pydantic (>=2.10.5,<3.0.0)",
    "langchain-openai (>=0.3.0,<0.4.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
